git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@742 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src_win/jdos/win/loader/winpe/LittleEndianFile.java
@@ -132,8 +132,8 @@ public class LittleEndianFile {
     }
 
     public final short readUnsignedByte() {
-        short result = Memory.mem_readb(address + pos);
+        int result = Memory.mem_readb(address + pos);
         pos+=1;
-        return result;
+        return (short)result;
     }
 }
